+++ realm/realm-library/src/objectServer/java/io/realm/SyncManager.java
@@ -21,6 +21,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.realm.annotations.Beta;
 import io.realm.internal.Keep;
 import io.realm.internal.network.AuthenticationServer;
@@ -43,7 +44,20 @@ import io.realm.log.RealmLog;
  */
 @Keep
 @Beta
+@SuppressFBWarnings("MS_CANNOT_BE_FINAL")
 public class SyncManager {
+
+    /**
+     * Debugging related options.
+     */
+    @SuppressFBWarnings("MS_SHOULD_BE_FINAL")
+    public static class Debug {
+        /**
+         * Set this to true to bypass checking if the device is offline before making HTTP requests.
+         */
+        public static boolean skipOnlineChecking = false;
+
+    }
 
     /**
      * APP ID sent to the Realm Object Server. Is automatically initialized to the package name for the app.