+++ realm/realm-library/src/objectServer/java/io/realm/permissions/PermissionChange.java
@@ -18,6 +18,7 @@ package io.realm.permissions;
 import java.util.Date;
 import java.util.UUID;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.realm.RealmObject;
 import io.realm.annotations.PrimaryKey;
 import io.realm.annotations.Required;
@@ -78,10 +79,12 @@ public class PermissionChange extends RealmObject {
         return id;
     }
 
+    @SuppressFBWarnings("EI_EXPOSE_REP")
     public Date getCreatedAt() {
         return createdAt;
     }
 
+    @SuppressFBWarnings("EI_EXPOSE_REP")
     public Date getUpdatedAt() {
         return updatedAt;
     }