+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.3.0
+
+### Object Server API Changes (In Beta)
+
+* Add a default `UserStore` based on the Realm Object Store (`ObjectStoreUserStore`).
+* Change the order of arguments to SyncCredentials.custom to match iOS: token, provider, userInfo
+* `SyncUser.all()` now returns Map instead of List.
+
 ## 2.2.3
 
 ### Bug fixes
@@ -37,6 +45,7 @@
 * All major public classes are now non-final. This is mostly a compromise to support Mockito. All protected fields/methods are still not considered part of the public API and can change without notice (#3869).
 * All Realm instances share a single notification daemon thread.
 * Fixed Java lint warnings with generated proxy classes (#2929).
+* Add 'like' predicate for String fields (#3752)
 
 ### Internal